requirements.txt
Flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
gunicorn==22.0.0
google-cloud-firestore==2.16.0
